Use paths and smaller window
Because of some unpredictable behaviour when not using absolute
paths, the command line tools have been moved to the tools dir
and are called called with tools/[meta]flac.exe. This way Windows
shouldn't choose a [meta]flac.exe somewhere else on the system.
This release uses a smaller window (the default 80) because of
Case's line break patch, which makes a wide console unneccesary.